Create server.go file to separate implementation of running http server from main.go file.
