+++ go.mod
@@ -3,6 +3,6 @@ module github.com/takumi616/go-english-vocabulary-api
 go 1.20
 
 require (
-	github.com/caarlos0/env/v6 v6.10.1 // indirect
-	golang.org/x/sync v0.6.0 // indirect
+	github.com/caarlos0/env/v6 v6.10.1
+	golang.org/x/sync v0.6.0
 )
